+++ src/intersection.rs
@@ -4,7 +4,7 @@ use material::Material;
 
 pub struct Intersection {
   pub position: Vector,
-  pub distance: f64,
+  pub distance: f32,
   pub normal: Vector,
   pub material: Arc<Material + Send + Sync>,
 }
