+++ src/sample.rs
@@ -1,4 +1,4 @@
 pub struct Sample<T> {
   pub value: T,
-  pub pdf: f64,
+  pub pdf: f32,
 }
